fix: input and filenames
